+++ scripts/macros/generic.mjs
@@ -22,3 +22,4 @@ export {reduceMaxHP} from './generic/reduceMaxHP.mjs';
 export {pushOnFail} from './generic/pushOnFail.mjs';
 export {grapple} from './generic/grapple.mjs';
 export {teleport} from './generic/teleport.mjs';
+export {rerollCheckWithBonus, rerollSaveWithBonus, rerollSkillWithBonus, rerollToolWithBonus} from './generic/rerollWithBonus.mjs';