+++ scripts/macros/generic/movementAnimation.mjs
@@ -1,12 +1,12 @@
-import {automationUtils, documentUtils, effectUtils} from '../../proxy.mjs';
+import {automationUtils, documentUtils, effectUtils, genericUtils, tokenUtils} from '../../proxy.mjs';
 async function use({document, workflow}) {
+    const bonus = workflow.activity.range.value || workflow.rangeDetails.range;
+    const range = bonus || tokenUtils.movementSpent(workflow.token.document, true);
+    if (!range) return genericUtils.notify('CHRISPREMADES.Macros.Generic.MovementAnimation.NoMovement', {type: 'warn'});
     const {animation: selectLocationsAnimation, options: selectLocationsOptions} = automationUtils.getResolvedAnimation(document, 'selectLocationsAnimation', {source: 'chris-premades', identifier: 'movementAnimation'});
     if (!selectLocationsAnimation) return;
     const {animation: moveAnimation} = automationUtils.getResolvedAnimation(document, 'moveAnimation', {source: 'chris-premades', identifier: 'movementAnimation'});
     if (!moveAnimation) return;
-    const movementType = automationUtils.getGenericConfigValue(document, 'chris-premades', 'movementAnimation', 'movementType');
-    const movement = workflow.actor.system.attributes.movement;
-    const range = movement[movementType] || movement.max;
     const positions = await selectLocationsAnimation.macros?.select(workflow.token.document, range, selectLocationsOptions);
     if (!positions?.length) return;
     const movementEffectId = automationUtils.getGenericConfigValue(document, 'chris-premades', 'movementAnimation', 'movementEffect');
@@ -55,13 +55,6 @@ export const movementAnimation = {
             type: 'selectAnimation',
             inputs: ['token', 'positions'],
             label: 'CHRISPREMADES.Macros.Generic.MovementAnimation.MoveAnimation',
-            hint: ''
-        },
-        movementType: {
-            default: 'max',
-            type: 'select',
-            get options() { return [{value: 'max', label: _loc('CHRISPREMADES.Macros.Generic.MovementAnimation.Fastest')}, ...Object.entries(CONFIG.DND5E.movementTypes).filter(([, data]) => !data.hidden).map(([value, data]) => ({value, label: _loc(data.label)}))]; },
-            label: 'CHRISPREMADES.Macros.Generic.MovementAnimation.MovementType',
             hint: ''
         },
         movementEffect: {